WIP on storage.

+++ msc.h
@@ -139,8 +139,8 @@ void MSC_ReadTrig(void);
 void MSC_ClassRequest(void);
 void MSC_SetConfig(void);
 
-void MSC_ReadMedia(uint32_t addr, uint32_t size, uint8_t *buffer);
-void MSC_WriteMedia(uint32_t addr, uint32_t size, uint8_t *buffer);
+void MSC_ReadMedia(uint64_t addr, uint64_t size, uint8_t *buffer);
+void MSC_WriteMedia(uint64_t addr, uint64_t size, uint8_t *buffer);
 
 /*-------------------------------------------------------------*/
 void MSC_AckCmd(void);
